Integrated core of 3.1 changes back into open stream with enterprise level signatures, Ldap, Oracle, and some validation plugins removed.

+++ app/models/task_context.rb
@@ -219,7 +219,7 @@ class TaskContext < ActiveRecord::Base
     if value.blank? or (value==DataElement::BLANK and cell.is_a?(TaskReference))
       item[:value] = ""
       item[:info] = "deleted #{cell.dom_id}"
-      cell.destroy
+      cell.destroy unless cell.new_record?
       return item
     end
     #
